+++ CHANGES.rst
@@ -13,6 +13,8 @@ Fixes:
 
 - Use zope.interface decorator.
   [gforcada]
+- Correctly log exception if viewlet rendering failed.
+  [jensens]
 
 
 2.0.9 (2015-09-07)